+++ desktop/cmp/clipboard/ClipboardButton.ts
@@ -10,10 +10,18 @@ import '@xh/hoist/desktop/register';
 import {Icon} from '@xh/hoist/icon';
 import {withDefault} from '@xh/hoist/utils/js';
 import copy from 'clipboard-copy';
+import {isString} from 'lodash';
 
 export interface ClipboardButtonProps extends ButtonProps {
     /** Function returning the text to copy. May be async. */
     getCopyText: () => string | Promise<string>;
+
+    /**
+     * Message to be displayed in a toast should the copy operation fail, or `true` (default) to
+     * show a toast-based alert from `XH.handleException`. Spec `false` to fail silently.
+     */
+    errorMessage?: string | boolean;
+
     /** Message to be displayed in a toast when copy is complete. */
     successMessage?: string;
 }
@@ -24,24 +32,35 @@ export interface ClipboardButtonProps extends ButtonProps {
 export const [ClipboardButton, clipboardButton] = hoistCmp.withFactory<ClipboardButtonProps>({
     displayName: 'ClipboardButton',
     model: false,
+
     render(props) {
-        let {icon, onClick, text, getCopyText, successMessage, ...rest} = props;
+        let {icon, onClick, text, getCopyText, errorMessage, successMessage, ...rest} = props;
+        errorMessage = withDefault(props.errorMessage, true);
 
         if (!onClick) {
             onClick = async () => {
-                const {successMessage, getCopyText} = props;
-
                 try {
                     const text = await getCopyText();
                     await copy(text);
                     if (successMessage) {
                         XH.toast({
-                            message: successMessage,
-                            icon: Icon.clipboard()
+                            icon: Icon.clipboard(),
+                            message: successMessage
                         });
                     }
                 } catch (e) {
-                    XH.handleException(e, {showAlert: false});
+                    if (isString(errorMessage)) {
+                        XH.dangerToast({
+                            icon: Icon.clipboard(),
+                            message: errorMessage
+                        });
+                    } else {
+                        XH.handleException(e, {
+                            title: 'Error copying to clipboard',
+                            showAlert: !!errorMessage,
+                            alertType: 'toast'
+                        });
+                    }
                 }
             };
         }